+++ llamazure/rbac/integration_test.py
@@ -1,4 +1,5 @@
 """Integration tests for roles"""
+
 import logging
 import os
 from typing import Any