+++ llamazure/tf/models.py
@@ -3,7 +3,14 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from collections import defaultdict
 from dataclasses import dataclass
-from typing import Union
+from typing import Generic, TypeVar, Union
+
+T = TypeVar("T")
+
+
+@dataclass(frozen=True)
+class TFRenderOpt(Generic[T]):
+	value: T
 
 
 class TFResource(ABC):