+++ llamazure/tf/network.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass, field
 from enum import Enum
 from typing import Generic, TypeVar
 
-from llamazure.tf.models import AnyTFResource, TFList, TFResource, _pluralise
+from llamazure.tf.models import AnyTFResource, TFList, TFRenderOpt, TFResource, _pluralise
 
 T = TypeVar("T")
 
@@ -24,6 +24,14 @@ class Counter(Generic[T]):
 		return v
 
 
+class OptNSGTotal(TFRenderOpt[bool]):
+	"""
+	Whether the rules for this NSG are total (there are no others).
+
+	Set this to `False` to allow for other `azurerm_network_security_rule` rules.
+	"""
+
+
 @dataclass
 class NSG(TFResource):
 	"""An azurerm_network_security_group resource"""
@@ -34,28 +42,40 @@ class NSG(TFResource):
 	rules: list[NSGRule]
 	tags: dict[str, str] = field(default_factory=dict)
 
+	opt_total: OptNSGTotal = OptNSGTotal(False)
+
 	@property
 	def t(self) -> str:  # type: ignore[override]
 		return "azurerm_network_security_group"
 
 	def render(self) -> dict:
 		"""Render for tf-json"""
+		if self.opt_total.value:
+			counter: Counter[NSGRule.Direction] = Counter(initial_value=100)
+
+			security_rules = [rule.render_as_block(counter.incr(rule.direction)) for rule in self.rules]
+		else:
+			security_rules = None
+
 		return {
 			"name": self.name,
 			"resource_group_name": self.rg,
 			"location": self.location,
-			"security_rule": None,
+			"security_rule": security_rules,
 			"tags": self.tags,
 		}
 
 	def subresources(self) -> list[TFResource]:
+		if self.opt_total.value:
+			return []
+
 		counter: Counter[NSGRule.Direction] = Counter(initial_value=100)
 
 		return [
 			AnyTFResource(
 				name="%s-%s" % (self.name, rule.name),
 				t="azurerm_network_security_rule",
-				props=rule.render("${%s.%s.name}" % (self.t, self.name), self.rg, counter.incr(rule.direction)),
+				props=rule.render_as_subresources("${%s.%s.name}" % (self.t, self.name), self.rg, counter.incr(rule.direction)),
 			)
 			for rule in self.rules
 		]
@@ -82,17 +102,25 @@ class NSGRule:
 	class Access(Enum):
 		"""Access type"""
 
-		Allow: str = "Allow"
-		Deny: str = "Deny"
+		Allow = "Allow"
+		Deny = "Deny"
 
 	class Direction(Enum):
 		"""Direction type"""
 
-		Inbound: str = "Inbound"
-		Outbound: str = "Outbound"
+		Inbound = "Inbound"
+		Outbound = "Outbound"
 
-	def render(self, nsg_name: str, rg: str, priority: int):
+	def render_as_subresources(self, nsg_name: str, rg: str, priority: int):
 		"""Render for tf-json"""
+		return {
+			**self.render_as_block(priority),
+			"resource_group_name": rg,
+			"network_security_group_name": nsg_name,
+		}
+
+	def render_as_block(self, priority: int):
+		"""Render for tf-json as a block on an `azurerm_network_security_group`"""
 		return {
 			"name": self.name,
 			"description": self.description,
@@ -104,8 +132,6 @@ class NSGRule:
 			"source_application_security_group_ids": self.src_sgids,
 			"destination_application_security_group_ids": self.dst_sgids,
 			"access": self.access.value,
-			"priority": priority,
 			"direction": self.direction.value,
-			"resource_group_name": rg,
-			"network_security_group_name": nsg_name,
+			"priority": priority,
 		}