+++ llamazure/azrest/changelog.md
@@ -1,5 +1,14 @@
 # 0
 
+## 0.4
+
+## 0.4.0
+
+- fix: improve inference of dicts 
+- feature: longpoll
+- feature: batch api can retry failed subtasks
+- feature: handle ratelimit retry-after (currently only in batches) 
+
 ## 0.3
 
 ### 0.3.0